refactor: blocks-handler
